WifiInstaller: remove the installation file
Previously, the installation file deletion was done by the parsing
function WifiManager#buildWifiConfig.  This results in a security
vulnerability with the parsing function since the caller can use
that function to delete arbitrary files.

The underlying API used by WifiManager#buildWifiConfig is updated
to not perform the file deletion.  So as the caller of that API,
we are responsible for deleting the installation file.

Bug: 33178389
Test: Verify passpoint configuration installation works using shamu
Change-Id: I3b88347c86dcb213033b5aa76e7e19a5524bee05
(cherry picked from commit bfd17d2ab2be44f9827bcbb4d57833698813f79b)